requirements.txt
asgiref
boto==2.49.0
boto3==1.20.22
botocore==1.23.22
certifi==2021.10.8
charset-normalizer==2.0.9
Django
django-filter==21.1
django-phonenumber-field==6.0.0
django-storages==1.12.3
djangorestframework==3.12.4
idna==3.3
importlib-metadata==4.8.2
jmespath==0.10.0
Markdown==3.3.6
phonenumbers==8.12.37
Pillow==8.4.0
psycopg2
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse
typing-extensions
urllib3==1.26.7
zipp==3.6.0
